Force use of trait when specifying job type
By default if a job type wasn't specified, the first job's type would be
inferred and used. This isn't correct, the job type generic should
*always* be a `dyn ___Marker` type which would allow all jobs within it
to be used, instead of just the single job struct. Force user to specify
a `dyn ___Marker` type using the `JobTypeMarker` trait, which is only
implemented on the dyn type, not its concrete implementations.